requirements.txt
google-api-python-client>=2.86.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pandas>=1.5.0
scikit-learn>=1.2.0
python-dateutil>=2.8.2
transformers>=4.28.0
torch>=2.0.0
nltk>=3.8.1
python-dotenv>=1.0.0
networkx>=3.1
plotly>=5.14.0
spacy>=3.5.0
PyQt6>=6.4.0